build: Add LD_PRELOAD lib as part of activation
To solve current error about tcmalloc or a freed pointer.
For example: `src/tcmalloc.cc:333] Attempt to free invalid pointer 0x5565a1917f00`
we need to add `LD_PRELOAD` as env variable before running any of the python
library. So we added that in the pixi env activation for now as
described in https://github.com/STEllAR-GROUP/phylanx/wiki/Phylanx-Issues-and-Answers.