add benchmarks (#25)
* feat: add hpyx.future.then() continuation

* feat: add init benchmarking suite

* feat: add hpx dot1d, tests, and benchmark

* build: deleted & re-created pixi.lock to resolve merge conflicts

* build: using system allocator to avoid tcmalloc build error on linux

* build: linux build changes for thread safety

* build: Updating host dependencies

Moving mimalloc, gperftools, and asio as host dependencies to try to solve tmalloc issue

* build: Update lockfile and libboost vers

Updating the lockfile with host-dependencies,
and reverting libboost-headers version
to match what's currently on main

* style: fix python linter build errors

* style: pre-commit fixes

* build: Add LD_PRELOAD lib as part of activation

To solve current error about tcmalloc or a freed pointer.
For example: `src/tcmalloc.cc:333] Attempt to free invalid pointer 0x5565a1917f00`
we need to add `LD_PRELOAD` as env variable before running any of the python
library. So we added that in the pixi env activation for now as
described in https://github.com/STEllAR-GROUP/phylanx/wiki/Phylanx-Issues-and-Answers.

* build: fix python linter errors

* style: pre-commit fixes

* build: disabled mypy linting for tests project folder to work around a false positive

* build: disable ruff linting for false positive on tests folder

* build: disabling ruff linting for the specific file rather than tests globally

* build: need mypy linting disabled on tests folder for separate false positive

---------

Co-authored-by: ayushnag <35325113+ayushnag@users.noreply.github.com>
Co-authored-by: Don Setiawan <landungs@uw.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>